Split the nnodely files
