+++ nnodely/exporter/__init__.py
@@ -1,3 +1,3 @@
 from nnodely.exporter.export import save_model, load_model, export_python_model
-from nnodely.exporter.exporter import Exporter
+from nnodely.exporter.emptyexporter import EmptyExporter
 from nnodely.exporter.standardexporter import StandardExporter
