+++ nnodely/exporter/standardexporter.py
@@ -1,6 +1,6 @@
 import os, torch
 
-from nnodely.exporter.exporter import Exporter
+from nnodely.exporter.emptyexporter import EmptyExporter
 from nnodely.exporter.reporter import Reporter
 from nnodely.exporter.export import save_model, load_model, export_python_model, export_pythononnx_model, export_onnx_model, import_python_model, import_onnx_model, onnx_inference
 from nnodely.utils import check
@@ -8,7 +8,7 @@ from nnodely.utils import check
 from nnodely.logger import logging, nnLogger
 log = nnLogger(__name__, logging.CRITICAL)
 
-class StandardExporter(Exporter):
+class StandardExporter(EmptyExporter):
     def __init__(self, workspace=None, visualizer=None, save_history=False):
         super().__init__(workspace, visualizer, save_history)
